Move ConsulSessionManager to consulutil.SessionManager
The ConsulSessionManager type had more to do with Consul than with the other
types in the "kp" package, so this commit moves it to the "consulutil" package.
