+++ bin/p2-rctl-server/main.go
@@ -14,6 +14,7 @@ import (
 
 	"github.com/square/p2/pkg/health/checker"
 	"github.com/square/p2/pkg/kp"
+	"github.com/square/p2/pkg/kp/consulutil"
 	"github.com/square/p2/pkg/kp/flags"
 	"github.com/square/p2/pkg/kp/rcstore"
 	"github.com/square/p2/pkg/kp/rollstore"
@@ -89,7 +90,7 @@ func main() {
 
 	// Start acquiring sessions
 	sessions := make(chan string)
-	go kp.ConsulSessionManager(api.SessionEntry{
+	go consulutil.SessionManager(api.SessionEntry{
 		Name:      SessionName(),
 		LockDelay: 5 * time.Second,
 		Behavior:  api.SessionBehaviorDelete,
